Added more options to quickEEmbed and createDeleteCollector
deleteOriginalMessage: by default is true and just deletes the message of the user calling the command and the message it self.

deleteCollectorOwnerId: used to check who can press the delete button, usually the delete collector can figure this out on its own but for certain cases where it cant you can set it manually.